+++ tests/__init__.py
@@ -2,5 +2,3 @@ import sys
 from os.path import join, dirname
 
 sys.path.insert(0, join(dirname(__file__), '..'))
-
-from . import test_coinaddr